feat: separate typescript files
