+++ src/types/components/buttonGroup.ts
@@ -1,0 +1,10 @@
+export type iButtonGroupProps = {
+  id?: string;
+  outline?: boolean;
+  className?: string;
+  children: React.ReactNode;
+  testId?: string;
+  onClick?: () => void;
+  onMouseEnter?: () => void;
+  onMouseLeave?: () => void;
+};
